+++ src/org/lukasj/idea/torquescript/asset/MaterialAssetImportPanel.kt
@@ -3,25 +3,28 @@ package org.lukasj.idea.torquescript.asset
 import com.intellij.openapi.project.Project
 import com.intellij.ui.dsl.builder.Panel
 import com.intellij.ui.dsl.builder.bindText
+import com.intellij.ui.dsl.builder.enableIf
 import org.lukasj.idea.torquescript.taml.MaterialAsset
 import java.nio.file.Path
 
 class MaterialAssetImportPanel(val project: Project, val asset: MaterialAsset) : AssetImportPanel {
     override fun insertPanel(panel: Panel) =
         panel.group("Material") {
-            row("Script File") {
-                textFieldWithBrowseButton { asset.assetFile.resolve(it.path).toString() }
+            row("Map To Material Name") {
+                textField()
                     .bindText(
-                        getter = { asset.scriptFile ?: "" },
-                        setter = { asset.scriptFile = asset.assetFile.parent.relativize(Path.of(it)).toString() }
+                        getter = { asset.mapTo ?: "" },
+                        setter = { asset.mapTo = it }
                     )
+                    .enabled(asset.mapTo == null)
             }
             row("Material Definition Name") {
                 textField()
                     .bindText(
-                        getter = { asset.materialDefinitionName ?: "" },
+                        getter = { asset.assetName ?: "" },
                         setter = { asset.materialDefinitionName = it }
                     )
+                    .enabled(false)
             }
         }
 }